Handle the FEATURES=noinfo case in emacs-infodir-rebuild; don't create a spurious Info dir file when called with no files in postrm phase. Package-Manager: portage-2.2_rc15/cvs/Linux 2.6.27-gentoo-r4 i686

+++ app-editors/emacs-cvs/emacs-cvs-23.0.9999.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2008 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/app-editors/emacs-cvs/emacs-cvs-23.0.9999.ebuild,v 1.13 2008/11/19 06:41:08 ulm Exp $
+# $Header: /var/cvsroot/gentoo-x86/app-editors/emacs-cvs/emacs-cvs-23.0.9999.ebuild,v 1.14 2008/11/26 21:27:00 ulm Exp $
 
 ECVS_AUTH="pserver"
 ECVS_SERVER="cvs.savannah.gnu.org:/sources/emacs"
@@ -239,13 +239,14 @@ emacs-infodir-rebuild() {
 	# INFOPATH, which is not guaranteed. So we rebuild it ourselves.
 
 	local infodir=/usr/share/info/${EMACS_SUFFIX} f
+	[ -d "${ROOT}"${infodir} ] || return	# may occur with FEATURES=noinfo
 	einfo "Regenerating Info directory index in ${infodir} ..."
 	rm -f "${ROOT}"${infodir}/dir{,.*}
 	for f in "${ROOT}"${infodir}/*.info*; do
-		[[ ${f##*/} == *[0-9].info* ]] \
-			|| install-info --info-dir="${ROOT}"${infodir} "${f}" &>/dev/null
+		[[ ${f##*/} != *[0-9].info* && -e ${f} ]] \
+			&& install-info --info-dir="${ROOT}"${infodir} "${f}" &>/dev/null
 	done
-	rmdir "${ROOT}"${infodir} 2>/dev/null # remove dir if it is empty
+	rmdir "${ROOT}"${infodir} 2>/dev/null	# remove dir if it is empty
 	echo
 }
